refactor: move more types into seperate directories
