+++ src/extension/features/system/types/IState.ts
@@ -2,6 +2,7 @@
 import type { ILogger } from '@common/types';
 
 interface IState {
+  deviceID: string | null;
   logger: ILogger;
   online: boolean;
 }
