+++ src/extension/features/system/utils/getInitialState.ts
@@ -6,6 +6,7 @@ import type { IState } from '../types';
 
 export default function getInitialState(): IState {
   return {
+    deviceID: null,
     logger: createLogger(__ENV__ === 'development' ? 'debug' : 'error'),
     online: true,
   };
