+++ src/extension/selectors/layout/index.ts
@@ -1,3 +1,4 @@
 export { default as useSelectConfirmModal } from './useSelectConfirmModal';
+export { default as useSelectDeviceID } from './useSelectDeviceID';
 export { default as useSelectScanQRCodeModal } from './useSelectScanQRCodeModal';
 export { default as useSelectSideBar } from './useSelectSideBar';
