+++ src/extension/types/accounts/IAddAccountCompleteResult.ts
