+++ src/extension/types/accounts/index.ts
@@ -2,3 +2,6 @@ export type { default as IAccount } from './IAccount';
 export type { default as IAccountInformation } from './IAccountInformation';
 export type { default as IAccountTransactions } from './IAccountTransactions';
 export type { default as IAccountWithExtendedProps } from './IAccountWithExtendedProps';
+export type { default as IActiveAccountDetails } from './IActiveAccountDetails';
+export type { default as IAddAccountCompleteResult } from './IAddAccountCompleteResult';
+export type { default as IInitializeAccountOptions } from './IInitializeAccountOptions';
