+++ src/extension/types/algod/IAlgorandAssetHolding.ts
