+++ src/extension/types/algod/IAlgorandPendingTransactionResponse.ts
@@ -1,5 +1,5 @@
 // types
-import IAlgorandTransaction from './IAlgorandTransaction';
+import type IAlgorandTransaction from './IAlgorandTransaction';
 
 interface IAlgorandPendingTransactionResponse {
   ['application-index']?: bigint;
