+++ src/extension/types/algod/IAlgorandStateSchema.ts
