+++ src/extension/types/algod/IAlgorandTealValue.ts
