+++ src/extension/types/i18n/IResourceLanguage.ts
