+++ src/extension/types/i18n/index.ts
@@ -1,0 +1,1 @@
+export type { default as IResourceLanguage } from './IResourceLanguage';
