+++ src/extension/types/index.ts
@@ -1,68 +1,31 @@
-export type { default as IActiveAccountDetails } from './IActiveAccountDetails';
-export type { default as IAddAccountCompleteResult } from './IAddAccountCompleteResult';
-export type { default as IAddAccountPageProps } from './IAddAccountPageProps';
-export type { default as IAlgorandAccountInformation } from './IAlgorandAccountInformation';
-export type { default as IAlgorandAccountTransaction } from './IAlgorandAccountTransaction';
-export type { default as IAlgorandApplication } from './IAlgorandApplication';
-export type { default as IAlgorandApplicationParams } from './IAlgorandApplicationParams';
-export type { default as IAlgorandApplicationTransaction } from './IAlgorandApplicationTransaction';
-export type { default as IAlgorandAsset } from './IAlgorandAsset';
-export type { default as IAlgorandAssetHolding } from './IAlgorandAssetHolding';
-export type { default as IAlgorandAssetConfigTransaction } from './IAlgorandAssetConfigTransaction';
-export type { default as IAlgorandAssetFreezeTransaction } from './IAlgorandAssetFreezeTransaction';
-export type { default as IAlgorandAssetParams } from './IAlgorandAssetParams';
-export type { default as IAlgorandAssetTransferTransaction } from './IAlgorandAssetTransferTransaction';
-export type { default as IAlgorandKeyRegistrationTransaction } from './IAlgorandKeyRegistrationTransaction';
-export type { default as IAlgorandPaymentTransaction } from './IAlgorandPaymentTransaction';
-export type { default as IAlgorandPendingTransactionResponse } from './IAlgorandPendingTransactionResponse';
-export type { default as IAlgorandSearchApplicationsResult } from './IAlgorandSearchApplicationsResult';
-export type { default as IAlgorandSearchAssetsResult } from './IAlgorandSearchAssetsResult';
-export type { default as IAlgorandStateSchema } from './IAlgorandStateSchema';
-export type { default as IAlgorandTealKeyValue } from './IAlgorandTealKeyValue';
-export type { default as IAlgorandTealValue } from './IAlgorandTealValue';
-export type { default as IAlgorandTransaction } from './IAlgorandTransaction';
-export type { default as IAlgorandTransactionParams } from './IAlgorandTransactionParams';
-export type { default as IAppProps } from './IAppProps';
-export type { default as IAppThunkDispatch } from './IAppThunkDispatch';
-export type { default as IAppThunkDispatchReturn } from './IAppThunkDispatchReturn';
-export type { default as IAppWindow } from './IAppWindow';
-export type { default as IAsyncThunkConfigWithRejectValue } from './IAsyncThunkConfigWithRejectValue';
-export type { default as IBaseActionMeta } from './IBaseActionMeta';
-export type { default as IBaseAsyncThunkConfig } from './IBaseAsyncThunkConfig';
-export type { default as IBaseTransaction } from './transactions/IBaseTransaction';
-export type { default as ICondensedProps } from './ICondensedProps';
 export type { default as IDecodedJwt } from './IDecodedJwt';
 export type { default as IDecodedJwtHeader } from './IDecodedJwtHeader';
 export type { default as IDecodedJwtPayload } from './IDecodedJwtPayload';
 export type { default as IFetchAssetWithDelayOptions } from './IFetchAssetWithDelayOptions';
-export type { default as IFulfilledActionMeta } from './IFulfilledActionMeta';
-export type { default as IInitializeAccountOptions } from './IInitializeAccountOptions';
-export type { default as IPasswordLock } from './IPasswordLock';
-export type { default as IPasswordTag } from './IPasswordTag';
-export type { default as IPendingActionMeta } from './IPendingActionMeta';
-export type { default as IPrivateKey } from './IPrivateKey';
-export type { default as IRegistrationAddAccountCompleteResult } from './IRegistrationAddAccountCompleteResult';
-export type { default as IRejectedActionMeta } from './IRejectedActionMeta';
-export type { default as IResourceLanguage } from './IResourceLanguage';
-export type { default as IScanQRCodeModalContentProps } from './IScanQRCodeModalContentProps';
-export type { default as ISession } from './ISession';
-export type { default as ISettings } from './settings/ISettings';
-export type { default as IStorageItemTypes } from './IStorageItemTypes';
 export type { default as ITinyManAssetResponse } from './ITinyManAssetResponse';
 export type { default as IUpdateAssetInformationByIdOptions } from './IUpdateAssetInformationByIdOptions';
 export type { default as IVestigeFiAssetResponse } from './IVestigeFiAssetResponse';
 export type { default as IWalletConnectSessionMetadata } from './IWalletConnectSessionMetadata';
 export * from './accounts';
+export * from './algod';
 export * from './arc-0072';
 export * from './arc-0200';
 export * from './arc-0300';
 export * from './asset-holdings';
 export * from './assets';
 export * from './events';
+export * from './i18n';
+export * from './layout';
 export * from './modals';
 export * from './networks';
 export * from './notifications';
+export * from './password-lock';
+export * from './private-key';
+export * from './redux';
+export * from './registration';
+export * from './sessions';
 export * from './settings';
 export * from './states';
+export * from './storage';
 export * from './styles';
 export * from './transactions';
