+++ src/extension/types/layout/IAddAccountPageProps.ts
@@ -1,5 +1,5 @@
 // types
-import IAddAccountCompleteResult from './IAddAccountCompleteResult';
+import type IAddAccountCompleteResult from '../accounts/IAddAccountCompleteResult';
 
 interface IProps {
   onComplete: (result: IAddAccountCompleteResult) => Promise<void>;
