+++ src/extension/types/layout/ICondensedProps.ts
