+++ src/extension/types/layout/IScanQRCodeModalContentProps.ts
