+++ src/extension/types/layout/index.ts
@@ -1,0 +1,5 @@
+export type { default as IAddAccountPageProps } from './IAddAccountPageProps';
+export type { default as IAppProps } from './IAppProps';
+export type { default as IAppWindow } from './IAppWindow';
+export type { default as ICondensedProps } from './ICondensedProps';
+export type { default as IScanQRCodeModalContentProps } from './IScanQRCodeModalContentProps';
