+++ src/extension/types/password-lock/IPasswordLock.ts
