+++ src/extension/types/password-lock/index.ts
@@ -1,0 +1,1 @@
+export type { default as IPasswordLock } from './IPasswordLock';
