+++ src/extension/types/private-key/index.ts
@@ -1,0 +1,2 @@
+export type { default as IPasswordTag } from './IPasswordTag';
+export type { default as IPrivateKey } from './IPrivateKey';
