+++ src/extension/types/redux/IBaseAsyncThunkConfig.ts
@@ -1,6 +1,6 @@
 // types
-import type IMainRootState from './states/IMainRootState';
-import type IRegistrationRootState from './states/IRegistrationRootState';
+import type IMainRootState from '../states/IMainRootState';
+import type IRegistrationRootState from '../states/IRegistrationRootState';
 
 interface IBaseAsyncThunkConfig<
   State = IMainRootState | IRegistrationRootState
