+++ src/extension/types/redux/IFulfilledActionMeta.ts
@@ -1,5 +1,5 @@
 // types
-import IBaseActionMeta from './IBaseActionMeta';
+import type IBaseActionMeta from './IBaseActionMeta';
 
 type IFulfilledActionMeta<Arg> = IBaseActionMeta<Arg, 'fulfilled'>;
