+++ src/extension/types/registration/index.ts
@@ -1,0 +1,1 @@
+export type { default as IRegistrationAddAccountCompleteResult } from './IRegistrationAddAccountCompleteResult';
