+++ src/extension/types/sessions/index.ts
@@ -1,0 +1,1 @@
+export type { default as ISession } from './ISession';
