+++ src/extension/types/storage/index.ts
@@ -1,0 +1,1 @@
+export type { default as IStorageItemTypes } from './IStorageItemTypes';
